feat(winui): capability-gated features + hardened scanner JSON contract
- ScannerService: GetCapabilitiesAsync() probes CLI version + per-subcommand
  --help and gates newer flags (sort-by/search/only-duplicates, category-totals,
  drop-relative, backfill-categories, prune-file-cache, prune-disk-space,
  progress-json). Flags default to supported and only clear on positive
  absence, so a version/help-format mismatch preserves behavior.
- StreamEvent.cs: StreamComplete.FileTypes int -> long (Rust emits u64 byte
  totals; int overflow threw and the streaming reader swallowed it, dropping
  the final scan result). ScanResult.FileTypes was already long.
- Tests: ScannerContractTests.cs pins the snake_case JSON contract (8 passed).

+++ gui-winui/SpaceAnalyzer/Models/StreamEvent.cs
@@ -81,8 +81,11 @@ public class StreamComplete
     [JsonPropertyName("duration_secs")]
     public double DurationSecs { get; set; }
 
+    // Long (not int): the Rust CLI emits per-type byte totals as u64. An int
+    // here would overflow on large directories and throw during JSON parsing,
+    // which the streaming reader swallows — silently dropping the final result.
     [JsonPropertyName("file_types")]
-    public Dictionary<string, int> FileTypes { get; set; } = new();
+    public Dictionary<string, long> FileTypes { get; set; } = new();
 
     [JsonPropertyName("extension_sizes")]
     public Dictionary<string, ulong> ExtensionSizes { get; set; } = new();
